% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cruciani.R
\docType{data}
\name{Cruciani}
\alias{Cruciani}
\title{A data frame of 3 Cruciani properties for 20 amino acids.}
\format{A data frame with 20 rows and 3 columns representing amino acids and Cruciani properties:
\describe{
  \item{}{Names of amino acids}
  \item{PP1}{Polarity}
  \item{PP2}{Hydrophobicity}
  \item{PP3}{H-bonding}
  ...
}}
\usage{
Cruciani
}
\description{
A data frame of 3 Cruciani properties for 20 amino acids.
}
\keyword{datasets}
