% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAASub.R
\name{getAASub}
\alias{getAASub}
\title{Get amino acid substitutions from multiple sequence alignment}
\usage{
getAASub(fileLoc)
}
\arguments{
\item{fileLoc}{exact location of multiple sequence alignment file in "FASTA" format}
}
\value{
Returns a list of two data frames 1.Single substitutions 2.Multiple substitutions
}
\description{
Get amino acid substitutions from multiple sequence alignment
}
\examples{
file = system.file("extdata", "linB_Prot_ali.fasta", package = "aaSEA")
getAASub(fileLoc = file)
}
