% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorPropChange.R
\name{getCorPropChange}
\alias{getCorPropChange}
\title{Get property changes associated with correlated substitutions}
\usage{
getCorPropChange(corSubFile, propertyDF = "Cruciani",
  propertyIndex = 1)
}
\arguments{
\item{corSubFile}{A dataframe with two columns i.e. Pos1 and Pos2. This file is generated by getTopSub}

\item{propertyDF}{Choose one of Cruciani, Fasgai, Kidera or AAindex based amino acid properties}

\item{propertyIndex}{The amino acid property to analyse. It is row number in propertyDF data frame}
}
\value{
A data frame of 12 columns consisting data regarding Wild type and mutant resides and their positions alogn with their properties and net changes between Pos1 and Pos2
}
\description{
Get property changes associated with correlated substitutions
}
\examples{
getCorPropChange(corSubFile = corSubFile, propertyDF = "Cruciani", propertyIndex = 1)
}
