% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorNet.R
\name{plotCorNet}
\alias{plotCorNet}
\title{simple interactive network diagram of correlated substitution pairs}
\usage{
plotCorNet(corSubDF)
}
\arguments{
\item{corSubDF}{A data frame with two columns i.e. Pos1 and Pos2. This file is generated by 'getTopSub' function.}
}
\value{
An interactive network diagram of correlated substitution pairs.
}
\description{
simple interactive network diagram of correlated substitution pairs
}
\examples{
corSub <- corSubFile
plotCorNet(corSubDF = corSub)
}
