% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorSubChanges.R
\name{plotCorSubChanges}
\alias{plotCorSubChanges}
\title{Plot co-evolving sites with selected property correlations}
\usage{
plotCorSubChanges(corSitePropChange)
}
\arguments{
\item{corSitePropChange}{A data frame of coevolving sites and associated property changes obtained by "getPropCorr" function with selected MSA matrix (selMat) and desired property selected from property data frame and property index.}
}
\value{
Returns an interactive heat map of significant sites with selected property correlations
}
\description{
Plot co-evolving sites with selected property correlations
}
\examples{
fileLocation <- system.file("extdata", "corSitePropChangeDF.rda", package = "aaSEA")
corSitePropChange <- readRDS(fileLocation)
plotCorSubChanges(corSitePropChange = corSitePropChange)
}
