% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiSubChange.R
\name{plotMultiSubChange}
\alias{plotMultiSubChange}
\title{Plots heat map of multiple substitution associated changes per site}
\usage{
plotMultiSubChange(multiSubChangeDF)
}
\arguments{
\item{multiSubChangeDF}{A data frame of multiple amino acid substitutions per site calculated with 'getAASub' and associated property changes obtained by calling 'getPropChange' function}
}
\value{
An interactive heat map of multiple substitution assocaiated changes per site
}
\description{
Plots heat map of multiple substitution associated changes per site
}
\examples{
multiSubChangeLoc <- system.file("extdata", "multiSubChange.rda", package = "aaSEA")
multiSubChange <- readRDS(multiSubChangeLoc)
plotMultiSubChange(multiSubChangeDF = multiSubChange )
}
