% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSingleSubChange.R
\name{plotSingleSubChange}
\alias{plotSingleSubChange}
\title{Plots single substitution change histogram}
\usage{
plotSingleSubChange(singleSubChangeDF)
}
\arguments{
\item{singleSubChangeDF}{A data frame of single amino acid substitutions per site calculated with 'getAASub' and associated property changes obtained by calling 'getPropChange' function}
}
\value{
An interactive histogram representing amino acid substitution associaited change
}
\description{
Plots single substitution change histogram
}
\examples{
singleSubChangeLoc <- system.file("extdata", "singleSubChange.rda", package = "aaSEA")
singleSubChange <- readRDS(singleSubChangeLoc)
plotSingleSubChange(singleSubChangeDF = singleSubChange)
}
