% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSeaShiny.R
\name{startSeaShiny}
\alias{startSeaShiny}
\title{Title Enables to start and run the app}
\usage{
startSeaShiny()
}
\value{
a shiny app will be launched in browser
}
\description{
Title Enables to start and run the app
}
\examples{
if(interactive()){
startSeaShiny()
}
}
